+++ src/drawer/index.ts
@@ -1,55 +1,15 @@
-import { createApp, App, h, ref, nextTick } from 'vue';
-import vueDrawer from './drawer.vue';
-import { WithInstallType } from '../shared';
-import { TdDrawerProps } from './type';
+import _Drawer from './drawer.vue';
+import { withInstall, WithInstallType } from '../shared';
 
 import './style';
 
-type DrawerOptions = Omit<TdDrawerProps, 'attach'>;
+import { TdDrawerProps } from './type';
 
-const Drawer = (options: DrawerOptions) => {
-  const root = document.createElement('div');
-  document.body.appendChild(root);
-  const visible = ref(false);
-  const props = ref({});
-  const destroyOnClose = ref(false);
-
-  createApp(() =>
-    h(vueDrawer, { ...options, visible: visible.value, destroyOnClose: destroyOnClose.value, ...props.value }),
-  ).mount(root);
-
-  const handler = {
-    destroy() {
-      destroyOnClose.value = true;
-      nextTick(() => {
-        visible.value = false;
-        document.body.removeChild(root);
-      });
-    },
-    hide() {
-      visible.value = false;
-    },
-    show() {
-      visible.value = true;
-    },
-    update(options: DrawerOptions) {
-      props.value = options;
-    },
-  };
-
-  return handler;
-};
-
-Drawer.install = (app: App, name = '') => {
-  app.component(name || vueDrawer.name, vueDrawer);
-
-  // 添加插件入口
-  app.config.globalProperties.$drawer = Drawer;
-  app.provide('$drawer', Drawer);
-};
-
-const _Drawer: WithInstallType<typeof Drawer> = Drawer;
-
-export default _Drawer;
+export type DividerProps = TdDrawerProps;
 
 export * from './type';
+export * from './plugin';
+export { default as DrawerPlugin } from './plugin';
+
+const Drawer: WithInstallType<typeof _Drawer> = withInstall(_Drawer);
+export default Drawer;
