+++ src/plugins.ts
@@ -1,3 +1,4 @@
 export { DialogPlugin } from './dialog';
 export { MessagePlugin } from './message';
 export { ToastPlugin } from './toast';
+export { DrawerPlugin } from './drawer';
